% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_responses.R
\name{get_time}
\alias{get_time}
\title{Convert timestamp in milliseconds to POSIXct format}
\usage{
get_time(timestamp)
}
\arguments{
\item{timestamp}{A numeric or character vector containing the timestamp in milliseconds.
This will be converted into seconds before being used to generate the corresponding POSIXct object.}
}
\value{
A POSIXct object representing the timestamp in UTC.
}
\description{
This function converts a timestamp in milliseconds (Unix time) to POSIXct format, which
can be used to represent dates and times in R. The timestamp is assumed to be in UTC.
}
\examples{
timestamp <- 1620000000000
converted_time <- get_time(timestamp)
print(converted_time)

}
